requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
camelot-py==1.0.9
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.122.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
Levenshtein==0.27.3
numpy==2.2.6
openai==2.8.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pillow==12.0.0
pluggy==1.6.0
portalocker==3.2.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==5.9.0
pypdfium2==5.1.0
pytesseract==0.3.13
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
